+++ packages/magnitude-core/src/types.ts
@@ -1,5 +1,5 @@
 import { FailureDescriptor } from "./common";
-import { TestCaseError } from "./errors";
+//import { TestCaseError } from "./errors";
 import { Ingredient } from "./recipe/types";
 
 export interface TestDataEntry {