+++ packages/magnitude-test/example/tests/magnitude.config.ts
@@ -1,5 +1,5 @@
 import { type MagnitudeConfig } from 'magnitude-test';
 
 export default {
-    url: "localhost:5173"
+    url: "https://magnitodo.com"
 } satisfies MagnitudeConfig;