+++ packages/magnitude-test/src/renderer/testRenderer.ts
@@ -171,7 +171,8 @@ export class TestCaseRenderer {
             // TODO: flesh out failure descriptor (via failure breakdown prompt) and show better here
             if (this.state.result && !this.state.result.passed) {
                 lines.push(chalk.redBright(`\nProblem:`));
-                lines.push(` ${this.state.result.failure.description}`);
+                lines.push(` ${this.state.result.failure}`);
+                //lines.push(` ${this.state.result.failure.description}`);
                 // const severity = problem.getSeverity();
                 // lines.push(` ${this.getSeverityDescriptor(severity)}: ${problem.getTitle()} `);
                 // lines.push(` ${magnitudeBlue('Expected')}: ${problem.getExpectedResult()}`);