+++ packages/magnitude-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # magnitude-core
 
+## 0.0.7
+
+### Patch Changes
+
+- improved check conversions, failure classification
+
 ## 0.0.6
 
 ### Patch Changes