+++ packages/magnitude-remote/CHANGELOG.md
@@ -1,5 +1,12 @@
 # magnitude-remote
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - magnitude-core@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes